pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "satyadrishti"
version = "0.1.0"
description = "Multimodal Deepfake & Coercion Detection Network — Edge-Deployed Digital Guardian"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.10"
authors = [
    {name = "Satya Drishti Team"},
]

dependencies = [
    # ── Core ML ──
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.38.0",
    "peft>=0.9.0",
    "accelerate>=0.27.0",
    "datasets>=2.18.0",
    "safetensors>=0.4.0",
    # ── Audio Processing ──
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    # ── Video / Vision ──
    "opencv-python>=4.9.0",
    "mediapipe>=0.10.9",
    "pillow>=10.2.0",
    # ── Edge Optimization ──
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    # ── Data & Utils ──
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    # ── Live Capture ──
    "pyaudio>=0.2.14",
    "websockets>=12.0",
    # ── Web Service ──
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "reportlab>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
    "ipykernel>=6.29.0",
]

[tool.setuptools.packages.find]
include = ["engine*", "server*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 120
target-version = ["py310"]
